requirements-server-ai.txt
Django==5.2.8
gunicorn==23.0.0
openai>=1.0.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
pillow-heif>=0.18.0
